+++ src/main/java/roomescape/dto/MemberReservationResponse.java
@@ -1,0 +1,24 @@
+package roomescape.dto;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+import roomescape.domain.Reservation;
+
+public record MemberReservationResponse(
+        Long reservationId,
+        ThemeResponse theme,
+        LocalDate date,
+        LocalTime time,
+        String status
+) {
+
+    public static MemberReservationResponse from(Reservation reservation) {
+        return  new MemberReservationResponse(
+                reservation.getId(),
+                ThemeResponse.from(reservation.getTheme()),
+                reservation.getDate(),
+                reservation.getReservationTime().getStartAt(),
+                "예약");
+    }
+}
